chore(logger): remove "export prodLogger" in logger
