+++ packages/logger/src/settings/browser.js
@@ -5,7 +5,7 @@
  * @flow
  */
 
-import type { settingsType } from '../logger';
+import type { settingsType } from '..';
 
 const { log, error, warn, info } = console;
