+++ packages/logger/src/settings/node.js
@@ -2,7 +2,7 @@
 
 import { mockChoice } from '@cat-org/utils';
 
-import type { settingsType } from '../logger';
+import type { settingsType } from '..';
 
 const { log, error, warn, info } = console;
